A few changes to the new repo organization: - removed field_utils.py (functions were moved elsewhere) - moved vcf-related functionalities to vcf.py (I envision this will become the place for all VCF needs)
