+++ gnomad/utils/vcf.py
@@ -1,0 +1,72 @@
+import hail as hl
+from typing import List
+import logging
+
+logging.basicConfig(format="%(levelname)s (%(name)s %(lineno)s): %(message)s")
+logger = logging.getLogger(__name__)
+logger.setLevel(logging.INFO)
+
+INFO_VCF_AS_PIPE_DELIMITED_FIELDS = ['AS_QUALapprox', 'AS_VarDP', 'AS_MQ_DP', 'AS_RAW_MQ', 'AS_SB_TABLE']
+
+def ht_to_vcf_mt(
+        info_ht: hl.Table,
+        pipe_delimited_annotations  : List[str] = INFO_VCF_AS_PIPE_DELIMITED_FIELDS
+ ) -> hl.MatrixTable:
+    """
+    Creates a MT ready for vcf export from a HT. In particular, the following conversions are done:
+    - All int64 are coerced to int32
+    - Fields specified by `pipe_delimited_annotations` will be converted from arrays to pipe-delimited strings
+
+    .. note::
+
+        The MT returned has no cols.
+
+    :param info_ht: Input HT
+    :param pipe_delimited_annotations: List of info fields (they must be fields of the ht.info Struct)
+    :return: MatrixTable ready for VCF export
+    """
+
+    def get_pipe_expr(array_expr: hl.expr.ArrayExpression) -> hl.expr.StringExpression:
+        return hl.delimit(array_expr.map(lambda x: hl.or_else(hl.str(x), "")), "|")
+
+    # Make sure the HT is keyed by locus, alleles
+    info_ht = info_ht.key_by('locus', 'alleles')
+
+    # Convert int64 fields to int32 (int64 isn't supported by VCF)
+    for f, ft in info_ht.info.dtype.items():
+        if ft == hl.dtype('int64'):
+            logger.warning(f"Coercing field info.{f} from int64 to int32 for VCF output. Value will be capped at int32 max value.")
+            info_ht = info_ht.annotate(
+                info=info_ht.info.annotate(**{f: hl.int32(hl.min(2**31 - 1, info_ht.info[f]))})
+            )
+        elif ft == hl.dtype('array<int64>'):
+            logger.warning(f"Coercing field info.{f} from array<int64> to array<int32> for VCF output. Array values will be capped at int32 max value.")
+            info_ht = info_ht.annotate(
+                info=info_ht.info.annotate(**{f: info_ht.info[f].map(lambda x: hl.int32(hl.min(2**31 - 1, x)))})
+            )
+
+    info_expr = {}
+
+    # Make sure to pipe-delimit fields that need to.
+    # Note: the expr needs to be prefixed by "|" because GATK expect one value for the ref (always empty)
+    # Note2: this doesn't produce the correct annotation for AS_SB_TABLE, but it is overwritten below
+    for f in pipe_delimited_annotations:
+        if f in info_ht.info:
+            info_expr[f] = "|" + get_pipe_expr(info_ht.info[f])
+
+    # Flatten SB if it is an array of arrays
+    if 'SB' in info_ht.info and not isinstance(info_ht.info.SB, hl.expr.ArrayNumericExpression):
+        info_expr['SB'] = info_ht.info.SB[0].extend(info_ht.info.SB[1])
+
+    if 'AS_SB_TABLE' in info_ht.info:
+        info_expr['AS_SB_TABLE'] = get_pipe_expr(info_ht.info.AS_SB_TABLE.map(lambda x: hl.delimit(x, ",")))
+
+    # Annotate with new expression and add 's' empty string field required to cast HT to MT
+    info_ht = info_ht.annotate(
+        info=info_ht.info.annotate(**info_expr),
+        s=hl.null(hl.tstr)
+    )
+
+    # Create an MT with no cols so that we acn export to VCF
+    info_mt = info_ht.to_matrix_table_row_major(columns=['s'], entry_field_name='s')
+    return info_mt.filter_cols(False)
